// 2024/2024-05-clever-platform-backoffice/v1/src/context/domain/g05/g05-d01/g05-d01-p02-teacher-homework/local/type.ts
export enum Status {
  IN_USE = 'ใช้งาน',
  DRAFT = 'แบบร่าง',
  NOT_IN_USE = 'ไม่ใช้งาน',
}
export enum StatusHomework {
  DEU = 'การบ้านที่ต้องส่ง',
  UPCOMMING = 'การบ้านที่สั่งล่วงหน้า',
  PAST = 'การบ้านที่ผ่านมา',
}
// ...
export interface HomeworkListItem {
  homework_id: number;
  homework_name: string;
  subject_id: number;
  subject_name: string;
  lesson_id: number;
  lesson_name: string;
  started_at: string;
  due_at: string;
  closed_at: string;
  level_count: number;
  assign_target_list: string[];
  total_student_count: number;
  on_time_student_count: number;
  status?: 'enabled' | 'disabled';
  late_student_count: number;
  doing_student_count: number;
  not_start_student_count: number;
}
// ...
export interface YearOption {
  id: number;
  year_name: string;
}

export interface ClassOption {
  id: number;
  class_name: string;
}